+++ e2e-tests/endpoints/polkadot/index.ts
@@ -8,6 +8,7 @@ import block3892620 from './3892620.json';
 import block4092619 from './4092619.json';
 import block4392619 from './4392619.json';
 import block4947391 from './4947391.json';
+import block5705186 from './5705186.json';
 
 export const polkadotEndpoints = [
 	['/blocks/943438', JSON.stringify(block943438)], //v17
@@ -20,4 +21,5 @@ export const polkadotEndpoints = [
 	['/blocks/4092619', JSON.stringify(block4092619)], //v28
 	['/blocks/4392619', JSON.stringify(block4392619)], //v29
 	['/blocks/4947391', JSON.stringify(block4947391)], //v30
+	['/blocks/5705186', JSON.stringify(block5705186)], //v9050
 ];