+++ scripts/types.ts
@@ -1,0 +1,18 @@
+export enum StatusCode {
+	Success = '0',
+	Failed = '1',
+}
+
+export interface IChainConfig {
+	wsUrl: string;
+	JestProcOpts: IProcOpts;
+	SasStartOpts: IProcOpts;
+}
+
+// Process options
+export interface IProcOpts {
+	proc: string;
+	resolver: string;
+	resolverStartupErr?: string;
+	args: string[];
+}