+++ src/chains-config/dockPoSTestnetControllers.ts
@@ -1,0 +1,32 @@
+import { ControllerConfig } from '../types/chains-config';
+import { getBlockWeight } from './metadata-consts';
+
+/**
+ * Controllers for Dock's test network.
+ */
+export const dockTestnetControllers: ControllerConfig = {
+	controllers: [
+		'AccountsBalanceInfo',
+		'AccountsStakingInfo',
+		'AccountsStakingPayouts',
+		'Blocks',
+		'BlocksExtrinsics',
+		'NodeNetwork',
+		'NodeTransactionPool',
+		'NodeVersion',
+		'PalletsStakingProgress',
+		'PalletsStorage',
+		'RuntimeCode',
+		'RuntimeMetadata',
+		'RuntimeSpec',
+		'TransactionDryRun',
+		'TransactionFeeEstimate',
+		'TransactionMaterial',
+		'TransactionSubmit',
+	],
+	options: {
+		finalizes: true,
+		minCalcFeeRuntime: 26,
+		blockWeightStore: getBlockWeight('dock-pos-test-runtime'),
+	},
+};