+++ src/controllers/blocks/BlocksController.ts
@@ -1,4 +1,5 @@
 import { ApiPromise } from '@polkadot/api';
+import { BlockHash } from '@polkadot/types/interfaces';
 import { isHex } from '@polkadot/util';
 import { RequestHandler } from 'express';
 
@@ -10,6 +11,12 @@ interface ControllerOptions {
 	finalizes: boolean;
 	minCalcFeeRuntime: null | number;
 	blockWeightStore: {};
+}
+
+interface IFinalizationOpts {
+	hash: BlockHash;
+	omitFinalizedTag: boolean;
+	queryFinalizedHead: boolean;
 }
 
 /**
@@ -89,6 +96,8 @@ export default class BlocksController extends AbstractController<BlocksService> 
 		this.safeMountAsyncGetHandlers([
 			['/head', this.getLatestBlock],
 			['/:number', this.getBlockById],
+			['/head/header', this.getLatestBlockHeader],
+			['/:number/header', this.getBlockHeaderById],
 		]);
 	}
 
@@ -105,21 +114,9 @@ export default class BlocksController extends AbstractController<BlocksService> 
 		const eventDocsArg = eventDocs === 'true';
 		const extrinsicDocsArg = extrinsicDocs === 'true';
 
-		let hash, queryFinalizedHead, omitFinalizedTag;
-		if (!this.options.finalizes) {
-			// If the network chain doesn't finalize blocks, we dont want a finalized tag.
-			omitFinalizedTag = true;
-			queryFinalizedHead = false;
-			hash = (await this.api.rpc.chain.getHeader()).hash;
-		} else if (finalized === 'false') {
-			omitFinalizedTag = false;
-			queryFinalizedHead = true;
-			hash = (await this.api.rpc.chain.getHeader()).hash;
-		} else {
-			omitFinalizedTag = false;
-			queryFinalizedHead = false;
-			hash = await this.api.rpc.chain.getFinalizedHead();
-		}
+		const paramFinalized = finalized === 'true' ? true : false;
+		const { hash, queryFinalizedHead, omitFinalizedTag } =
+			await this.parseFinalizationOpts(this.options.finalizes, paramFinalized);
 
 		const options = {
 			eventDocs: eventDocsArg,
@@ -168,5 +165,78 @@ export default class BlocksController extends AbstractController<BlocksService> 
 			res,
 			await this.service.fetchBlock(hash, options)
 		);
+	};
+
+	/**
+	 * Return the Header of the identified block.
+	 *
+	 * @param req Express Request
+	 * @param res Express Response
+	 */
+	private getBlockHeaderById: RequestHandler<INumberParam> = async (
+		{ params: { number } },
+		res
+	): Promise<void> => {
+		const hash = await this.getHashForBlock(number);
+
+		BlocksController.sanitizedSend(
+			res,
+			await this.service.fetchBlockHeader(hash)
+		);
+	};
+
+	/**
+	 * Return the header of the latest block
+	 *
+	 * @param req Express Request
+	 * @param res Express Response
+	 */
+	private getLatestBlockHeader: RequestHandler = async (
+		{ query: { finalized } },
+		res
+	): Promise<void> => {
+		const paramFinalized = finalized !== 'false';
+
+		const hash = paramFinalized
+			? await this.api.rpc.chain.getFinalizedHead()
+			: undefined;
+
+		BlocksController.sanitizedSend(
+			res,
+			await this.service.fetchBlockHeader(hash)
+		);
+	};
+
+	/**
+	 * This also returns the hash for the block to query.
+	 *
+	 * @param optFinalizes
+	 * @param paramFinalized
+	 */
+	private parseFinalizationOpts = async (
+		optFinalizes: boolean,
+		paramFinalized: boolean
+	): Promise<IFinalizationOpts> => {
+		let hash, queryFinalizedHead, omitFinalizedTag;
+		if (!optFinalizes) {
+			// If the network chain doesn't finalize blocks, we dont want a finalized tag.
+			omitFinalizedTag = true;
+			queryFinalizedHead = false;
+			hash = (await this.api.rpc.chain.getHeader()).hash;
+		} else if (!paramFinalized) {
+			omitFinalizedTag = false;
+			queryFinalizedHead = true;
+			hash = (await this.api.rpc.chain.getHeader()).hash;
+		} else {
+			omitFinalizedTag = false;
+			queryFinalizedHead = false;
+			hash = await this.api.rpc.chain.getFinalizedHead();
+		}
+
+		return {
+			hash,
+			omitFinalizedTag,
+			queryFinalizedHead,
+		};
 	};
 }