+++ src/logging/Log.ts
@@ -1,24 +1,26 @@
 import { createLogger, Logger } from 'winston';
+import { ConsoleTransportInstance } from 'winston/lib/winston/transports';
 
 import { consoleTransport } from './transports';
-
-// Note: there is a `fileTransport` that gets added in main.
-const transports = [consoleTransport()];
 
 /**
  * Access a singleton winston.Logger that will be intialized on first use.
  */
 export class Log {
+	private static _transports: ConsoleTransportInstance[] | undefined;
 	private static _logger: Logger | undefined;
 	private static create(): Logger {
 		if (this._logger) {
 			return this._logger;
 		}
 
+		// Note: there is a `fileTransport` that gets added in main.
+		this._transports = [consoleTransport()];
+
 		this._logger = createLogger({
-			transports,
+			transports: this._transports,
 			exitOnError: false,
-			exceptionHandlers: transports,
+			exceptionHandlers: this._transports,
 		});
 
 		return this._logger;